Refactoring of file structure in webhook folder
