+++ src/blueprints/telegram_bot/webhook/commands/create_folder.py
@@ -1,27 +1,18 @@
 from flask import g
 
-from .....api import telegram, yandex
-from ..decorators import (
+from .....api import telegram
+from .common.responses import (
+    cancel_command
+)
+from .common.decorators import (
     yd_access_token_required,
     get_db_data
 )
-from ..responses import (
-    cancel_command
+from .common.api import (
+    create_folder,
+    YandexAPIError,
+    YandexAPIRequestError
 )
-
-
-class YandexAPIRequestError(Exception):
-    """
-    Some error during Yandex.Disk API request occured.
-    """
-    pass
-
-
-class YandexAPIError(Exception):
-    """
-    Error response from Yandex.Disk API.
-    """
-    pass
 
 
 @yd_access_token_required
@@ -81,75 +72,3 @@ def get_text(message: dict) -> str:
         message.get("caption") or
         ""
     )
-
-
-def is_error_response(data: dict) -> bool:
-    """
-    :returns: Yandex response contains error or not.
-    """
-    return ("error" in data)
-
-
-def create_error_text(data: dict) -> str:
-    """
-    Constructs error text from Yandex error response.
-    """
-    error_name = data["error"]
-    error_description = (
-        data.get("message") or
-        data.get("description") or
-        "?"
-    )
-
-    return (
-        "Yandex.Disk Error: "
-        f"{error_name} ({error_description})"
-    )
-
-
-def create_folder(access_token: str, folder_name: str) -> int:
-    """
-    Creates folder using Yandex API.
-
-    Yandex not able to create folder if some of
-    middle folders not exists. This method will try to create
-    each folder one by one, and ignore safe errors (if
-    already exists, for example) from all folder names
-    except last one.
-
-    :returns: Last (for last folder name) HTTP Status code.
-
-    :raises: YandexAPIRequestError
-    :raises: YandexAPIError
-    """
-    folders = [x for x in folder_name.split("/") if x]
-    folder_path = ""
-    last_status_code = 201  # root always created
-    allowed_errors = [409]
-
-    for folder in folders:
-        response = None
-        folder_path = f"{folder_path}/{folder}"
-
-        try:
-            response = yandex.create_folder(
-                access_token,
-                path=folder_path
-            )
-        except Exception as e:
-            raise YandexAPIRequestError(e)
-
-        last_status_code = response["HTTP_STATUS_CODE"]
-
-        if (
-            (last_status_code == 201) or
-            (not is_error_response(response)) or
-            (last_status_code in allowed_errors)
-        ):
-            continue
-
-        raise YandexAPIError(
-            create_error_text(response)
-        )
-
-    return last_status_code
