+++ src/blueprints/telegram_bot/webhook/commands/upload_photo.py
@@ -4,15 +4,15 @@ from typing import Union
 from flask import g, current_app
 
 from .....api import telegram, yandex
-from ..decorators import (
+from .common.decorators import (
     yd_access_token_required,
     get_db_data
 )
-from ..responses import (
+from .common.responses import (
     abort_command,
     cancel_command
 )
-from .create_folder import (
+from .common.api import (
     create_folder,
     YandexAPIRequestError,
     YandexAPIError
